// features/work-item-management/src/lib/work-item/components/work-items-board/model/work-item-board-filter.model.ts
import {
  WorkItemPriority,
  WorkItemStatus,
  WorkItemType,
} from "../../../model/work-item";

export interface WorkItemBoardFilter {
  search?: string;
  projectIds?: string[];
  workItemStatuses?: WorkItemStatus[];
  workItemPriority?: WorkItemPriority;
  workItemType?: WorkItemType;
  assignees?: string[];
  workItemCategoryContains?: string;
  dueDateFrom?: string;
  dueDateTo?: string;
  resolvedDateSince?: string;
  workItemCategories?: string[];
  objectIds?: string[];
}